+++ .github/workflows/ci-nightly.yml
@@ -244,6 +244,14 @@ jobs:
             os: ubuntu-latest
             files: >-
               tests/integration/test_slow_agni_aragog.py
+          # ---- shard: grid (dummy-backend parameter grid; a single
+          # module-scoped fixture runs the whole grid, ~27 min. Linux-only
+          # extended because it needs no macOS parity) ----
+          - shard: grid
+            tier: extended
+            os: ubuntu-latest
+            files: >-
+              tests/grid/test_grid.py
           # ---- shard: janus-inference (JANUS + BO + inference) ----
           - shard: janus-inference
             tier: critical
@@ -251,30 +259,32 @@ jobs:
             files: >-
               tests/integration/test_slow_janus_aragog.py
               tests/inference/test_bo_convergence.py
-              tests/inference/test_inference.py
+              tests/inference/test_inference_slow.py
           - shard: janus-inference
             tier: critical
             os: macos-latest
             files: >-
               tests/integration/test_slow_janus_aragog.py
               tests/inference/test_bo_convergence.py
-              tests/inference/test_inference.py
+              tests/inference/test_inference_slow.py
           # ---- shard: misc (config + smoke + dummy + helpers) ----
           - shard: misc
             tier: critical
             os: ubuntu-latest
             files: >-
-              tests/config/test_config_schema_invariants.py
               tests/integration/test_smoke_hypothesis.py
               tests/integration/test_integration_dummy.py
+              tests/integration/test_integration_std_config.py
+              tests/interior_struct/test_superliquidus_adiabat.py
               tests/helpers/test_smoke_invariants.py
           - shard: misc
             tier: critical
             os: macos-latest
             files: >-
-              tests/config/test_config_schema_invariants.py
               tests/integration/test_smoke_hypothesis.py
               tests/integration/test_integration_dummy.py
+              tests/integration/test_integration_std_config.py
+              tests/interior_struct/test_superliquidus_adiabat.py
               tests/helpers/test_smoke_invariants.py
     runs-on: ${{ matrix.os }}
     timeout-minutes: 210   # job cap for the slow-tier matrix, sized by its longest shard (zalmoxis-coupled/ubuntu, ~137 min plus ~11 min setup); above the 180 min per-test pytest-timeout so a hung single test trips pytest-timeout's per-test timer before the job is cancelled.
